fix: enhance perf of Text when fontSize changed in webgl (#1424)
* fix: enhance perf of Text when fontSize changed in webgl #1422

* chore: commit changeset

* fix: draw 1px sub-pixel line correctly in webgl #1425

* chore: commit changeset

* chore: add testcase for sub-pixel line #1425

* fix: convertToPath should account for Rect with undefined x/y

* chore: commit changeset

* fix: make FillMesh instanced to enhance perf #1359

* chore: commit changeset

* fix: make textBaseline in SVG the same with Canvas #1313

* chore: commit changeset

* fix: add a fixed offset for Text in webgl #1380

* chore: commit changeset

* fix: convertToPath should be compatible with empty coords

* chore: commit changeset
