+++ .changeset/cool-chefs-grin.md
@@ -1,0 +1,7 @@
+---
+'@antv/g-plugin-device-renderer': patch
+'@antv/g-shader-components': patch
+'@antv/g-lite': patch
+---
+
+Make FillMesh instanced to enhance perf.
