+++ .changeset/eight-beers-hide.md
@@ -1,0 +1,6 @@
+---
+'@antv/g-plugin-device-renderer': patch
+'@antv/g-lite': patch
+---
+
+ConvertToPath should account for Rect with undefined x/y.
