+++ .changeset/hungry-emus-invite.md
@@ -1,0 +1,8 @@
+---
+'@antv/g-plugin-device-renderer': patch
+'@antv/g-plugin-svg-renderer': patch
+'@antv/g-shader-components': patch
+'@antv/g-lite': patch
+---
+
+Make textBaseline in SVG the same with Canvas.
