+++ .changeset/itchy-balloons-yell.md
@@ -1,0 +1,8 @@
+---
+'@antv/g-plugin-device-renderer': patch
+'@antv/g-plugin-svg-renderer': patch
+'@antv/g-shader-components': patch
+'@antv/g-lite': patch
+---
+
+Add a fixed offset for Text.
