+++ .changeset/sharp-moose-leave.md
@@ -1,0 +1,8 @@
+---
+'@antv/g-plugin-device-renderer': patch
+'@antv/g-plugin-svg-renderer': patch
+'@antv/g-shader-components': patch
+'@antv/g-lite': patch
+---
+
+ConvertToPath should be compatible with empty coords.
