+++ .changeset/shiny-crabs-press.md
@@ -1,0 +1,5 @@
+---
+'@antv/g-plugin-device-renderer': patch
+---
+
+Draw 1px sub-pixel line correctly in webgl.
