+++ .changeset/young-donkeys-return.md
@@ -1,0 +1,5 @@
+---
+'@antv/g-plugin-device-renderer': patch
+---
+
+Enhance perf of Text when fontSize changed in webgl.
