+++ packages/g-plugin-device-renderer/src/meshes/index.ts
@@ -2,8 +2,7 @@ export * from './Instanced';
 export * from './SDF';
 export * from './InstancedLine';
 export * from './InstancedPath';
-export * from './Line';
-export * from './Fill';
+export * from './InstancedFill';
 export * from './Image';
 export * from './Text';
 export * from './Mesh';
