+++ packages/g-plugin-device-renderer/src/shader/instanced-line.frag
@@ -24,7 +24,12 @@ void main() {
         outputColor = u_Color;
     #endif
 
-    float blur = smoothstep(0.0, v_Distance.y, 1.0 - abs(v_Distance.x));
+    float blur;
+    if (v_Distance.y < 1.0) {
+      blur = smoothstep(0.0, v_Distance.y, 1.0 - abs(v_Distance.x));
+    } else {
+      blur = 1.0 / v_Distance.y;
+    }
     float u_dash_offset = v_Dash.y;
     float u_dash_array = v_Dash.z;
     float u_dash_ratio = v_Dash.w;
