+++ packages/g-plugin-device-renderer/src/shader/instanced-line.vert
@@ -30,6 +30,7 @@ void main() {
   } else {
     strokeWidth = u_StrokeWidth;
   }
+  float clampedStrokeWidth = max(strokeWidth, 1.0);
 
   float isBillboard = a_Dash.w;
   bool isPerspective = isPerspectiveMatrix(u_ProjectionMatrix);
@@ -43,8 +44,8 @@ void main() {
     vec2 screen1 = u_Viewport * (0.5 * clip1.xy / clip1.w + 0.5);
     vec2 xBasis = normalize(screen1 - screen0);
     vec2 yBasis = vec2(-xBasis.y, xBasis.x);
-    vec2 pt0 = screen0 + strokeWidth * (a_Position.x * xBasis + a_Position.y * yBasis);
-    vec2 pt1 = screen1 + strokeWidth * (a_Position.x * xBasis + a_Position.y * yBasis);
+    vec2 pt0 = screen0 + clampedStrokeWidth * (a_Position.x * xBasis + a_Position.y * yBasis);
+    vec2 pt1 = screen1 + clampedStrokeWidth * (a_Position.x * xBasis + a_Position.y * yBasis);
     vec2 pt = mix(pt0, pt1, a_Position.z);
     vec4 clip = mix(clip0, clip1, a_Position.z);
     gl_Position = vec4(clip.w * (2.0 * pt / u_Viewport - 1.0), clip.z, clip.w);
@@ -52,12 +53,12 @@ void main() {
     vec2 xBasis = a_PointB.xy - a_PointA.xy;
     vec2 yBasis = normalize(vec2(-xBasis.y, xBasis.x));
 
-    vec2 point = a_PointA.xy + xBasis * a_Position.x + yBasis * strokeWidth * a_Position.y;
+    vec2 point = a_PointA.xy + xBasis * a_Position.x + yBasis * clampedStrokeWidth * a_Position.y;
     point = point - u_Anchor.xy * abs(xBasis);
 
     // round & square
     if (a_Cap > 1.0) {
-      point += sign(a_Position.x - 0.5) * normalize(xBasis) * vec2(strokeWidth / 2.0);
+      point += sign(a_Position.x - 0.5) * normalize(xBasis) * vec2(clampedStrokeWidth / 2.0);
     }
     gl_Position = project(vec4(point, u_ZIndex, 1.0), u_ProjectionMatrix, u_ViewMatrix, u_ModelMatrix);
   }
