+++ packages/g-plugin-device-renderer/src/shader/mesh.vert
@@ -1,5 +1,7 @@
 #pragma glslify: import('@antv/g-shader-components/scene.both.glsl')
-#pragma glslify: import('@antv/g-shader-components/mesh.both.glsl')
+
+#pragma glslify: import('@antv/g-shader-components/batch.declaration.vert')
+#pragma glslify: project = require('@antv/g-shader-components/project.vert')
 
 layout(location = POSITION) in vec2 a_Position;
 #ifdef USE_UV
@@ -8,6 +10,8 @@ layout(location = POSITION) in vec2 a_Position;
 #endif
 
 void main() {
+  #pragma glslify: import('@antv/g-shader-components/batch.vert')
   #pragma glslify: import('@antv/g-shader-components/uv.vert')
-  gl_Position = u_ProjectionMatrix * u_ViewMatrix * u_ModelMatrix * vec4(a_Position, u_ZIndex, 1.0);
+
+  gl_Position = project(vec4(a_Position, u_ZIndex, 1.0), u_ProjectionMatrix, u_ViewMatrix, u_ModelMatrix);
 }
