+++ packages/g-plugin-device-renderer/src/shader/text.vert
@@ -22,13 +22,14 @@ void main() {
 
   bool isBillboard = a_StylePacked2.y > 0.5;
   float sizeAttenuation = a_StylePacked2.z;
-  vec2 offset = a_Offset * fontScale;
+  vec2 bufferOffset = vec2(0.7, 2.0);
+  vec2 offset = a_Offset * fontScale + bufferOffset;
 
   if (isBillboard) {
     #pragma glslify: import('@antv/g-shader-components/billboard.vert')
     v_GammaScale = 1.0;
   } else {
-    gl_Position = project(vec4(a_Offset * fontScale, u_ZIndex, 1.0), u_ProjectionMatrix, u_ViewMatrix, u_ModelMatrix);
+    gl_Position = project(vec4((a_Offset) * fontScale + bufferOffset, u_ZIndex, 1.0), u_ProjectionMatrix, u_ViewMatrix, u_ModelMatrix);
     v_GammaScale = gl_Position.w;
   }
 }
