+++ packages/g-plugin-svg-renderer/src/shapes/paths/Text.ts
@@ -63,8 +63,11 @@ export function updateTextElementAttribute(
 
     // Since `text-after-edge` is not a standard property value, we use `dy` instead.
     if (textBaseline === 'bottom' || textBaseline === 'top') {
-      $el.setAttribute('dominant-baseline', 'middle');
-      $el.setAttribute('dy', textBaseline === 'bottom' ? `-0.5em` : '0.5em');
+      $el.setAttribute('dominant-baseline', BASELINE_MAP['middle']);
+      $el.setAttribute(
+        'dy',
+        textBaseline === 'bottom' ? `-${height / 2}px` : `${height / 2}px`,
+      );
     }
 
     // <textPath> only support one line
