+++ site/examples/perf/animation/demo/meta.json
@@ -27,6 +27,20 @@
         "zh": "移动大量节点",
         "en": "Move nodes"
       }
+    },
+    {
+      "filename": "webgl-text.js",
+      "title": {
+        "zh": "在 WebGL 中更新 Text 的位置",
+        "en": "Update position of Text in WebGL"
+      }
+    },
+    {
+      "filename": "webgl-opacity.js",
+      "title": {
+        "zh": "在 WebGL 中更新 Opacity",
+        "en": "Update opacity of shapes in WebGL"
+      }
     }
   ]
 }
