+++ site/examples/shape/text/demo/meta.json
@@ -29,6 +29,13 @@
       "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*VAQWQbMdQkQAAAAAAAAAAAAADmJ7AQ/original"
     },
     {
+      "filename": "text-billboard.js",
+      "title": {
+        "zh": "文本公告牌效果",
+        "en": "Text Billboard"
+      }
+    },
+    {
       "filename": "web-font-loader.js",
       "title": {
         "zh": "加载字体",
